Corrected component naming in Form
