+++ superset-frontend/src/components/Form/index.tsx
@@ -16,62 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import React from 'react';
-import { Input } from 'antd';
-import { styled } from '@superset-ui/core';
-// import * as alertSVG from 'images/icons/alert.svg';
 import Form from './Form';
 import FormItem from './FormItem';
 import FormLabel from './FormLabel';
+import LabeledErrorBoundInput from './LabeledErrorBoundInput';
 
-// console.log('---alertString--- ', alertSVG);
-
-export interface LabeledErrorBoundInputProps {
-  label?: string;
-  name: string;
-  validationMethods:
-    | { onBlur: (value: any) => void }
-    | { onChange: (value: any) => void };
-  errorMessage: string | null;
-  helpText?: string;
-  value: string | number | readonly string[] | undefined;
-  required?: boolean;
-  placeholder?: string;
-  autoComplete?: string;
-  type?: string;
-  id?: string;
-  onChange?: any;
-}
-
-const StyledInput = styled(Input)`
-  margin: 8px 0;
-`;
-
-const LabeledErrorBoundInput = ({
-  label,
-  validationMethods,
-  errorMessage,
-  helpText,
-  required = false,
-  id,
-  ...props
-}: LabeledErrorBoundInputProps) => (
-  <>
-    <FormLabel htmlFor={id} required={required}>
-      {label}
-    </FormLabel>
-    <FormItem
-      validateTrigger={Object.keys(validationMethods)}
-      validateStatus={errorMessage ? 'error' : 'success'}
-      help={errorMessage || helpText}
-      hasFeedback={!!errorMessage}
-      {...props}
-    >
-      <StyledInput {...props} {...validationMethods} />
-    </FormItem>
-  </>
-);
-
-export default LabeledErrorBoundInput;
-
-export { Form, FormItem, FormLabel };
+export { Form, FormItem, FormLabel, LabeledErrorBoundInput };
